Merging changes from zw_misc to master.

+++ include/jonoondb_api/document_id_generator.h
@@ -12,7 +12,7 @@ class DocumentIDGenerator {
   DocumentIDGenerator(const DocumentIDGenerator&) = delete;
   DocumentIDGenerator(DocumentIDGenerator&&) = delete;
   DocumentIDGenerator& operator=(const DocumentIDGenerator&) = delete;
-  std::uint64_t ReserveID(std::uint32_t numOfIDsToReserve);
+  std::uint64_t ReserveID(std::uint64_t numOfIDsToReserve);
  private:
   std::atomic<std::uint64_t> m_currentID;
 };
